+++ HISTORY.rst
@@ -2,3 +2,4 @@
 ====================
 * Added overlay option [do3cc]
 * Sparse button gets hidden now when the feature is not being used [do3cc]
+* Added css class to nodes that got found while filtering [do3cc]